fix: walletconnect message sign, reload persistance
