+++ src/components/SignForm/SignForm.js
@@ -1,14 +1,8 @@
-import NETWORKS from '../../consts/networks.js'
 import MESSAGE_TYPES from '../../consts/messageTypes'
 
 import Tippy from '@tippyjs/react'
 
-import { init, useConnectWallet, useWallets } from '@web3-onboard/react'
-import walletConnectModule from '@web3-onboard/walletconnect'
-import ledgerModule from '@web3-onboard/ledger'
-import injectedModule from '@web3-onboard/injected-wallets'
-import trezorModule from '@web3-onboard/trezor'
-import gnosisModule from '@web3-onboard/gnosis'
+import { useConnectWallet, useSetChain, useWallets } from '@web3-onboard/react'
 
 import { ethers } from 'ethers'
 import { useCallback, useEffect, useState } from 'react'
@@ -24,56 +18,18 @@ import { getMessagePlaceholder, validateMessage } from '../../helpers/messages'
 import './SignForm.scss'
 import { MdIosShare } from 'react-icons/md'
 
-const WC_PROJECT_ID = 'd98522bddb36e73acae903da02b45fd1'
-
-const walletConnect = walletConnectModule({
-  projectId: WC_PROJECT_ID,
-  requiredChains: [],
-  optionalChains: NETWORKS.map((n) => n.chainId),
-  dappUrl: 'https://sigtool.ambire.com/',
-})
-
-const injected = injectedModule()
-const ledger = ledgerModule({
-  projectId: WC_PROJECT_ID,
-})
-const trezor = trezorModule({
-  appUrl: 'https://sigtool.ambire.com/',
-  email: 'contactus@ambire.com',
-})
-const gnosis = gnosisModule({ whitelistedDomains: [/./] })
-
-init({
-  wallets: [injected, walletConnect, trezor, ledger, gnosis],
-  chains: NETWORKS.map((n) => ({
-    id: ethers.utils.hexValue(n.chainId),
-    label: n.name,
-    rpcUrl: n.rpc,
-    token: n.token,
-  })),
-  appMetadata: {
-    name: 'Signature Validator',
-    icon: process.env.REACT_APP_SUBFOLDER_PATH + '/img/signature-validator-logo.png',
-    description: 'Signature Validator tool',
-    recommendedInjectedWallets: [{ name: 'MetaMask', url: 'https://metamask.io' }],
-  },
-  accountCenter: {
-    desktop: {
-      enabled: false,
-    },
-  },
-})
-
 const truncateAddress = (addr) => {
   return addr.substr(0, 4) + '...' + addr.substr(-4)
 }
 
+let provider
+
 const SignForm = ({ selectedForm, setShareModalLink }) => {
   const [{ wallet, connecting }, connect, disconnect] = useConnectWallet()
+  const [{ connectedChain }] = useSetChain()
 
   const connectedWallets = useWallets()
   const [connectedAccount, setConnectedAccount] = useState(null)
-  const [connectedChain, setConnectedChain] = useState(null)
 
   const [error, setError] = useState(null)
 
@@ -106,10 +62,12 @@ const SignForm = ({ selectedForm, setShareModalLink }) => {
   )
 
   useEffect(() => {
-    if (selectedForm !== 'sign') return
-    connect()
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [selectedForm])
+    if (!wallet?.provider) {
+      provider = null
+    } else {
+      provider = new ethers.providers.Web3Provider(wallet.provider, 'any')
+    }
+  }, [wallet])
 
   // wallet sign call
   const walletSign = useCallback(
@@ -119,26 +77,23 @@ const SignForm = ({ selectedForm, setShareModalLink }) => {
         return
       }
 
-      if (wallet.provider) {
-        const provider = new ethers.providers.Web3Provider(wallet.provider, 'any')
+      if (provider) {
         const signer = provider.getUncheckedSigner()
 
         if (messageType === 'humanMessage') {
           if (wallet.label === 'Gnosis Safe') {
-            return (await wallet.provider.sdk.txs.signMessage(hexlify(ethers.utils.toUtf8Bytes(message)))).safeTxHash
+            return (await provider.sdk.txs.signMessage(hexlify(ethers.utils.toUtf8Bytes(message)))).safeTxHash
           }
           return signer.signMessage(message)
         } else if (messageType === 'hexMessage') {
           if (wallet.label === 'Gnosis Safe') {
-            return (await wallet.provider.sdk.txs.signMessage(message)).safeTxHash
+            return (await provider.sdk.txs.signMessage(message)).safeTxHash
           }
           return signer.signMessage(arrayify(message))
         } else if (messageType === 'typedData') {
-          if (wallet.label === 'WalletConnect') {
-            return wallet.provider.connector.signTypedData([connectedAccount.address, message])
-          } else if (wallet.label === 'Gnosis Safe') {
+          if (wallet.label === 'Gnosis Safe') {
             return (
-              await wallet.provider.sdk.txs.signMessage({
+              await provider.sdk.txs.signMessage({
                 signType: 'eth_signTypedData_v4',
                 message,
               })
@@ -152,7 +107,7 @@ const SignForm = ({ selectedForm, setShareModalLink }) => {
         setError('Provider not found')
       }
     },
-    [wallet, connectedAccount]
+    [wallet]
   )
 
   // sign action
@@ -306,16 +261,15 @@ const SignForm = ({ selectedForm, setShareModalLink }) => {
 
   // only filter 1 main account
   useEffect(() => {
-    console.log(connectedWallets)
+    if (connecting) return
+
     if (!connectedWallets || connectedWallets.length === 0) {
       setConnectedAccount(null)
-      setConnectedChain(null)
       return
     }
     const firstWallet = connectedWallets[0]
     setConnectedAccount(firstWallet?.accounts[0])
-    setConnectedChain(firstWallet?.chains[0])
-  }, [connectedWallets])
+  }, [connectedWallets, connecting])
 
   useEffect(() => {
     setSignature(null)
